Move and rename example templates routes
