+++ app.js
@@ -24,7 +24,7 @@ const routes = require('./app/routes');
 const utils = require('./lib/utils');
 
 const prototypeAdminRoutes = require('./middleware/prototype-admin-routes');
-const exampleTemplatesRoutes = require('./docs/documentation_routes');
+const exampleTemplatesRoutes = require('./lib/example_templates_routes');
 
 // Set configuration variables
 const port = parseInt(process.env.PORT, 10) || config.port;
