+++ lib/example_templates_routes.js
